% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkr_tons_per_acre.R
\name{bkr_tons_per_acre}
\alias{bkr_tons_per_acre}
\title{Translate 18 square foot plot yields into tons per acre}
\usage{
bkr_tons_per_acre(x)
}
\description{
Translate 18 square foot plot yields into tons per acre
}
