% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datestamp_data.R
\name{datestamp_data}
\alias{datestamp_data}
\title{Insert a datestamp into a vector of data, probably data exported to breedbase}
\usage{
datestamp_data(x, my_date = "string")
}
\description{
Insert a datestamp into a vector of data, probably data exported to breedbase
}
