% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ccb_breedbase.R
\name{open_ccb_breedbase}
\alias{open_ccb_breedbase}
\title{configure and list trials for ccb BB, helpful precursor to pulling data}
\usage{
open_ccb_breedbase()
}
\description{
configure and list trials for ccb BB, helpful precursor to pulling data
}
