% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_study_data.R
\name{pull_study_data}
\alias{pull_study_data}
\title{Pull data from some study, simple helper function combining two QBMS functions}
\usage{
pull_study_data(x = "string")
}
\description{
Pull data from some study, simple helper function combining two QBMS functions
}
