% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_filename.R
\name{timestamp_filename}
\alias{timestamp_filename}
\title{Insert a timestamp into a string, likely a filename}
\usage{
timestamp_filename(my_File_name = "string")
}
\arguments{
\item{my_File_name}{A string, intended to be a filename (i.e. ends in .csv, .xlsx, etc.)}
}
\value{
a string containing a timestamp added. Timestamp is rounded to the nearest minute
}
\description{
Insert a timestamp into a string, likely a filename
}
\examples{
timestamp_filename("my_awesome_dataset.csv")
timestamp_filename("superduper.awesome_dataset.xlsx")
}
