% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkr_tons_per_acre.R
\name{bkr_tons_per_acre}
\alias{bkr_tons_per_acre}
\title{Yields in tons per acre assuming 18 ft^2 plot}
\usage{
bkr_tons_per_acre(x)
}
\arguments{
\item{x}{A number}
}
\value{
A number
}
\description{
Yields in tons per acre assuming 18 ft^2 plot
}
\examples{
bkr_tons_per_acre(c(2.25, 6))
}
