% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ccb_breedbase.R
\name{open_ccb_breedbase}
\alias{open_ccb_breedbase}
\title{configure and list trials for ccb BB, helpful precursor to pulling data}
\usage{
open_ccb_breedbase()
}
\arguments{
\item{Function}{takes no arguments}
}
\value{
A list of trials to be explored further via \code{QBMS}
}
\description{
configure and list trials for ccb BB, helpful precursor to pulling data
}
\examples{
open_ccb_breedbase()
QBMS::set_trial("2023")
QBMS::list_studies()
}
