% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_study_data.R
\name{pull_study_data}
\alias{pull_study_data}
\title{Pull data from some study, simple helper function combining two QBMS functions}
\usage{
pull_study_data(x = "string")
}
\arguments{
\item{x}{A string that is the name of a study available in Breedbase.}
}
\value{
data from some study
}
\description{
Pull data from some study, simple helper function combining two QBMS functions
}
\examples{
open_ccb_breedbase()
QBMS::set_trial("2023")
QBMS::list_studies()
pull_study_data("23WIWPALT")
lapply(c("23NYCCALT", "23NYHVALT"), pull_study_data)
}
