% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_filename.R
\name{timestamp_filename}
\alias{timestamp_filename}
\title{Insert a timestamp into a string, likely a filename}
\usage{
timestamp_filename(my_File_name = "string")
}
\description{
Insert a timestamp into a string, likely a filename
}
